// src/app/precisa/maestros/FormMaestro/model/maestro.ts
export class Maestro{

    constructor(){
    }
    Numero: number;
    IdTablaMaestro: number;
    CodigoTabla: string;
    Nombre: string;
    Descripcion: string;
    Version: string;
    Estado: number;
    UsuarioCreacion: string;
    FechaCreacion: Date;
    UsuarioModificacion: string;
    FechaModificacion: Date;
    EstadoDe: string;
}