// src/app/precisa/maestros/Empleados/model/empleadomast.ts
export class EmpleadoMast{

    constructor(){
        
    }
    IdEmpleado: number;
    Empleado: number;
    FechaInicioContrato: Date;
    FechaFinContrato: Date;
    FechaCese: Date;
    CompaniaSocio: string;
    Cargo: string;
    TipoTrabajador: string;
    Estado: string;
    DesCargo: string;
    codigo: number;
    TIPODOCUMENTO: string;
    Documento: string;
    DocumentoIdentidad: string;
    NombreCompleto: string;
    Telefono: string;
    Direccion: string;
    CorreoInterno: string;
    UltimoUsuario: string;
    UltimaFechaModif: Date;
    DesTipoTabajador: string;
    UneDescripcion: string;
    UneuNegocioId: number;
    DescEstado: string;
    DesDocumento:string;
    UsuarioCreacion?:       string;
    FechaCreacion?:         Date;
}