// src/app/precisa/maestros/Empleados/model/filtro.empleadomast.ts
export class filtroEmpleadoMast{

    constructor(){
    }

    UneuNegocioId:  number;
    CompaniaSocio:  string;
    IdEmpleado:     number;
    TipoDocumento:  string;
    Documento:      string;
    NombreCompleto: string;
    TipoTrabajador: string;
    Cargo:          string;
    Estado:         string;

}