// src/app/precisa/maestros/Empleados/model/iempleado.ts
export interface IEmpleado {
    NRO: number;
    TIPO_DOCUMENTO: string;
    DOCUMENTO_EMPLEADO: string;
    EMPLEADO: string;
    FECHA_INICIO: string;
    FECHA_FIN:string;
    FECHA_CESE: string;
    COMPAÑIA: string;
    CARGO: string;
    EMAIL: string;
    TELEFONO: string;
    DIRECCION: string;
    ESTADO: string;
  }
  